finished alerts back and front, select option alert done

+++ dis-app/src/main/java/it/akademija/models/UserData.java
@@ -53,8 +53,8 @@ public class UserData {
 		this.email = email;
 	}
 
-	public UserData(long id, String name, String surename, Integer personId, String address, String city,
-			Integer phoneNum, String email) {
+	public UserData(long id, String name, String surename, Long personId, String address, String city, Integer phoneNum,
+			String email) {
 		this.id = id;
 		this.name = name;
 		this.surename = surename;
